requirements.txt
antlr4-python3-runtime==4.8
apple-certifi==1.1.1
argcomplete==1.12.3
async-generator==1.10
boltproto==2.35.0
boto3==1.18.55
botocore==1.21.58
Cerberus==1.3.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.2
Cython==0.29.24
dataclasses==0.8
dpath==1.5.0
faiss-cpu==1.7.1.post2
filelock==3.3.0
gitdb==4.0.7
GitPython==3.1.18
grpcio==1.41.0
huggingface-hub==0.0.19
hydra-core==1.1.1
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.19.5
omegaconf==2.1.1
packaging==21.0
pandas==1.1.5
protobuf==3.17.3
pyotp==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
recertifi==1.1.1
regex==2021.10.8
requests==2.26.0
s3transfer==0.5.0
sacremoses==0.0.46
sentencepiece==0.1.96
six==1.16.0
smmap==4.0.0
tabulate==0.8.9
tokenizers==0.8.1rc1
torch==1.9.1
tqdm==4.62.3
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
watchdog==2.1.6
websockets==9.1
wget==3.2
wrapt==1.13.1
zipp==3.6.0
spacy
jsonlines
black
